+++ package.json
@@ -12,7 +12,7 @@
     "font-awesome": "^4.7.0",
     "husky": "^4.2.3",
     "jwt-decode": "^2.2.0",
-    "mdbreact": "^4.25.5",
+    "mdbreact": "^4.27.0",
     "node-sass": "^4.13.1",
     "prettier": "^2.0.2",
     "pretty-quick": "^2.0.1",
