+++ src/App.scss
@@ -1,6 +1,6 @@
 @import "./components/slider/slider.scss";
 @import "./components/layout/Navbar.scss";
-@import "./components/carousel/bookContentCarousel.scss";
+@import "./components/carousel/BookContentCarousel.scss";
 
 html,
 body {
@@ -27,7 +27,7 @@ body {
   margin: 5em 0 0 0;
   /* background-color: rgb(245, 245, 245); */
   border-radius: 0.5em;
-  padding: 20px;
+  padding: 10px;
   flex-basis: 100%;
   max-width: 75%;
 }
