+++ src/components/carousel/BookContentCarousel.scss
